use bump_version.sh in makedist.sh

+++ miniwiki/bump_version.sh
@@ -5,7 +5,9 @@
 version=$1
 sed 's/\(define("MW_VERSION", "\)[^"]\+/\1'$version'/' lib/miniwiki.php > lib/miniwiki.php.tmp
 mv lib/miniwiki.php.tmp lib/miniwiki.php
-sed 's/\(PROJECT_NUMBER.*=\).*$/\1 '$version'/' Doxyfile > Doxyfile.tmp
-mv Doxyfile.tmp Doxyfile
+if [ -e Doxyfile ]; then
+  sed 's/\(PROJECT_NUMBER.*=\).*$/\1 '$version'/' Doxyfile > Doxyfile.tmp
+  mv Doxyfile.tmp Doxyfile
+fi
 sed 's/\(miniWiki \).*\( (c)\)/\1'$version'\2/' README > README.tmp
 mv README.tmp README
